pyproject.toml
[project]
name = "cupel"
version = "0.1.64"
description = "separates precious LLMs from base LLMs. works with any OpenAI/Anthropic compatible API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "tolitius"},
]
keywords = ["llm", "eval", "benchmark", "leaderboard", "local-llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "rich>=13.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
]

[project.urls]
Homepage = "https://github.com/tolitius/cupel"
Repository = "https://github.com/tolitius/cupel"

[project.scripts]
cupel = "cupel.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cupel"]
